+++ src/main/java/javaProjects/HRMS/business/abstracts/JobAdvertisement/JobAdvertisementService.java
@@ -20,6 +20,6 @@ public interface JobAdvertisementService extends BaseService<JobAdvertisement, I
 	DataResult<List<JobAdvertisement>> getAllActiveByCompanyName(String companyName);
 	DataResult<List<JobAdvertisement>> getAllActiveAndConfirmed();
 	DataResult<List<JobAdvertisement>> getAllActiveAndConfirmedByPage(int pageNo,int pageSize);
-	Result setConfirmed(int jobAdvertisementId, Long systemEmployeeId);
+	Result setConfirmed(int jobAdvertisementId, Integer systemEmployeeId);
 
 }
