+++ src/main/java/javaProjects/HRMS/business/abstracts/Users/EmployerService.java
@@ -6,7 +6,7 @@ import javaProjects.HRMS.core.business.abstracts.BaseService;
 import javaProjects.HRMS.core.utilities.results.DataResult;
 import javaProjects.HRMS.entities.concretes.Users.Employer;
 
-public interface EmployerService extends BaseService<Employer, Long> {
+public interface EmployerService extends BaseService<Employer, Integer> {
 	
 	DataResult<Employer> getByEmail(String email);
 	void save(Employer employer);
