+++ src/main/java/javaProjects/HRMS/dataAccess/abstracts/Users/CandidateDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import javaProjects.HRMS.entities.concretes.Users.Candidate;
 
-public interface CandidateDao extends JpaRepository<Candidate, Long> {
+public interface CandidateDao extends JpaRepository<Candidate, Integer> {
 
 	Candidate getByIdentityNumber(String identityNumber);
 	Candidate getByEmail(String email);
